DO NOT APPLY: Enable QEMU_CAPS_BLOCKDEV if 'copy-on-read' is supported
Use the 'copy-on-read' block drive in qemu as a witness to enable the
blockdev functionality and adapt the test files for the new output.

Note that the blockjobs were NOT adapted yet so any blockjob will desync
the state of qemu. This patch needs to wait until blockjobs are fixed.
Message-Id: <2afa9727e874c1f24de97a84cab0ed6dbf6c3c69.1532533587.git.pkrempa@redhat.com>