networkGetDHCPLeases: Don't always report error if unable to read leases file
https://bugzilla.redhat.com/show_bug.cgi?id=1600468

If we are unable to read leases file (no matter what the reason
is), we return 0 - just like if there were no leases. However,
because we use virFileReadAll() an error is printed into the log.
Note that not all networks have leases file - only those for
which we start dnsmasq.

Signed-off-by: Michal Privoznik <mprivozn@redhat.com>
Message-Id: <520d46fdb1b65fa93e2a08d100e1f1461a900ba5.1532591717.git.mprivozn@redhat.com>
Reviewed-by: Erik Skultety <eskultet@redhat.com>