virtestmock: Track action
As advertised in the previous commit, we need the list of
accessed files to also contain action that caused the $path to
appear on the list. Not only this enables us to fine tune our
white list rules it also helps us to see why $path is reported.
For instance:

  /run/user/1000/libvirt/libvirt-sock: connect: qemuxml2argvtest: QEMU XML-2-ARGV net-vhostuser-multiq

Signed-off-by: Michal Privoznik <mprivozn@redhat.com>
Message-Id: <f72317c8a54c6eca1bf70b04e1d8b40792fb3357.1532704945.git.mprivozn@redhat.com>
Reviewed-by: John Ferlan <jferlan@redhat.com>