util: netdevmacvlan: define cleanup function using VIR_DEFINE_AUTOPTR_FUNC
Using the new VIR_DEFINE_AUTOPTR_FUNC macro defined in
src/util/viralloc.h, define a new wrapper around an existing
cleanup function which will be called when a variable declared
with VIR_AUTOPTR macro goes out of scope. Also, drop the redundant
viralloc.h include, since that has moved from the source module into
the header.

When a variable of type virNetlinkCallbackDataPtr is declared using
VIR_AUTOPTR, the function virNetlinkCallbackDataFree will be run
automatically on it when it goes out of scope.

This commit also adds an intermediate typedef for virNetlinkCallbackData
type for use with the cleanup macros.

Signed-off-by: Sukrit Bhatnagar <skrtbhtngr@gmail.com>
Message-Id: <1532800907-29115-15-git-send-email-skrtbhtngr@gmail.com>