tests: Unify data structure for vircaps2xmltest
If all we achieve is reducing the depth by one for a single
test case, the additional complexity (not to mention breaking
the principle of least surprise) is not worth it: let's use
simpler, more predictable code instead.

This basically reverts fec6e4c48c9c (with a few adjustments).

Signed-off-by: Andrea Bolognani <abologna@redhat.com>
Message-Id: <20180802164056.18218-1-abologna@redhat.com>
Reviewed-by: Martin Kletzander <mkletzan@redhat.com>