process: accept the lack of /proc/<pid> as valid process removal
There were cases where the process was gone (no /proc/<pid> entry anymore), but
kill with signal=0 still was able to reach the process.

This can happen if the kernel still cleans up resources.

In most common cases of this there would be a /proc/<pid> entry left with the
process in Zombie state until reaped (by init). But those cases usually resolve
rather quickly as init periodically will call wait to reap.

The more critical and confusing cases are those where the process is
gone from all that (not in /proc/<pid> anymore), but the kernel still
considers it reachable by kill with signal 0.

This is due to kill (2) only checking for "existence of a process ID" but not
the process itself.

This effect has mostly been seen when using plenty of SR-IOV resources in the
guest (which might explain the extra cleanup phase by the kernel) and to
resolve those issues libvirt will accept /proc/<pid> being gone as valid exit
as well (on top of signal 0 returning ESRCH).

Signed-off-by: Christian Ehrhardt <christian.ehrhardt@canonical.com>
Message-Id: <20180803062941.31641-3-christian.ehrhardt@canonical.com>