bhyve: fix process reconnect
To reconnect to the bhyve process after deamon restart, a process
VM's pid points to is checked to have proctitle equal to 'bhyve: $vmname'.
However, there could be a bug in bhyve(8) which prevents it from
setting proctitle, so process arguments will look like:

 ['/usr/sbin/bhyve', ..., 'vmname', NULL]

Fall back to checking this format if proctitle doesn't match as a
workaround for this bug.
Message-Id: <20180805155428.40866-1-bogorodskiy@gmail.com>