virlockspace: Introduce VIR_LOCK_SPACE_ACQUIRE_METADATA flag
This flag is going to be used to alter default behaviour of the
lock. Firstly, it means we will lock different offset in the file
(offset 1 instead of 0). Secondly, it means the lock acquiring
will actually wait for the lock to be set (compared to default
behaviour which is set or error out).

Signed-off-by: Michal Privoznik <mprivozn@redhat.com>
Message-Id: <7d81280d3e128ca98b1348c1e457dd8ce84c858f.1533821288.git.mprivozn@redhat.com>