domain_lock: Implement metadata locking
In order for our drivers to lock resources for metadata change we
need set of new APIs. Fortunately, we don't have to care about
every possible device a domain can have. We care only about those
which can live on a network filesystem and hence can be accessed
by multiple daemons at the same time. These devices are covered
in virDomainLockMetadataLock() and only a small fraction of
those can be hotplugged (covered in the rest of the introduced
APIs).

Signed-off-by: Michal Privoznik <mprivozn@redhat.com>
Message-Id: <3cc745a517b023b38d98930800840b60729a55a5.1533821288.git.mprivozn@redhat.com>