esx:Fix esxDomainGetMaxVcpus to return correct vcpus
Before this change, esxDomainGetMaxVcpus returned -1, which in turn
fails in libvirt. This commit reimplements esxDomainGetMaxVcpus instead
of calling esxDomainGetVcpusFlags. The implementation checks for
capability.maxSupportedVcpus, but as this one can be ommited in ESXi, we
also check for capability.maxHostSupportedVcpus. With this change,
virDomainSetVcpus, virDomainGetMaxVcpus and virDomainGetVcpusFlags and
returning correct values.

Signed-off-by: Marcos Paulo de Souza <marcos.souza.org@gmail.com>
Message-Id: <20180810035658.13555-3-marcos.souza.org@gmail.com>