storage: Properly terminate secrets
The virSecretGetSecretString() helper looks up a secret for given
pool and returns its value in @secret_value and its length in
@secret_value_size. However, the trailing '\0' is not included in
either of the variables. This is because usually the value of the
secret is passed to some encoder (usually base64 encoder) where
the trailing zero must not be accounted for.

However, in two places we actually want the string as we don't
process is any further.

Signed-off-by: Michal Privoznik <mprivozn@redhat.com>
Message-Id: <e78549665ba4ce6cc2f999ac16b64a70d6ef3bb8.1534149729.git.mprivozn@redhat.com>
Reviewed-by: J�n Tomko <jtomko@redhat.com>