storage: Don't hold storage pool locked during wipeVol
Currently the way virStorageVolWipe() works is it looks up
pool containing given volume and hold it locked throughout while
API execution. This is suboptimal because wiping a volume means
writing data to it which can take ages. And if the whole pool is
locked during that operation no other API can be issued (nor
pool-list).

Signed-off-by: Michal Privoznik <mprivozn@redhat.com>
Message-Id: <e4c0c13ab8e42fa3a2571fc6f5ce303dd8fa4363.1534157377.git.mprivozn@redhat.com>