domain_lock: Implement metadata locking
In order for our drivers to lock resources for metadata change we
need set of new APIs. Fortunately, we don't have to care about
every possible device a domain can have. We care only about those
which can live on a network filesystem and hence can be accessed
by multiple daemons at the same time. These devices are covered
in virDomainLockMetadataLock() and only a small fraction of
those can be hotplugged (covered in the rest of the introduced
APIs).

Signed-off-by: Michal Privoznik <mprivozn@redhat.com>
Message-Id: <1ffbcb28175f23b547a1be4c448abfce07a1f49e.1534245398.git.mprivozn@redhat.com>