virstorageobj: Check for duplicates from virStoragePoolObjAssignDef
Even though we do some checking is not as thorough as it should
be. We already have virStoragePoolObjIsDuplicate but the way we
use it is a typical TOCTOU. Imagine two threads trying to define
two pools with the same name but different UUIDs. With the
current code neither of them finds a duplicate and thus proceed
to virStoragePoolObjAssignDef where only names are compared.
Therefore both threads succeed which is obviously wrong.

We should check for duplicates where we care for them.

Signed-off-by: Michal Privoznik <mprivozn@redhat.com>
Message-Id: <cfeec09bf3875735ae907606a21f0d58e7eb4540.1534766849.git.mprivozn@redhat.com>
Reviewed-by: John Ferlan <jferlan@redhat.com>