qemu: Add check for unpriv sgio for SCSI generic host device
Check if the hostdev has set the sgio filtered/unfiltered and handle
appropriately.

This restores functionality removed by commit id 'ce346623' to remove
sgio support for the SCSI generic host device in the formerly named
method qemuCheckSharedDevice.

For most kernels the result of this operation is a no-op; however, for
those that do support it the check is necessary.

Note that this patch fixes a bug found in the reverted change where
if the qemuGetDeviceUnprivSGIO returned either 0 or 1 in 'val', the
'disk' would be dereferenced; however, since a hostdev didn't have
one - that dereference would have caused a segfault. That was fixed
by commit id 'cd01d2ad5' using virDomainDiskGetSource. Instead these
changes use the hostdev's address value.

Signed-off-by: John Ferlan <jferlan@redhat.com>
Message-Id: <20180823155002.19288-3-jferlan@redhat.com>