virLockManagerLockDaemonAddResource: Switch to cleanup label rather than error
This will help in future expansions of the code when it is be
harder to track if @newName and/or @newLockspace were already
allocated or not and thus whether it is safe to 'return' or we
need to 'goto error'. By using the 'cleanup' label those two
cases merge into a single one.

Signed-off-by: Michal Privoznik <mprivozn@redhat.com>
Message-Id: <1c240add261aceb100e25e43e560e9f5416eabfe.1535356707.git.mprivozn@redhat.com>
Reviewed-by: John Ferlan <jferlan@redhat.com>