lock_daemon_dispatch: Check for ownerPid rather than ownerId
At the beginning of each dispatch function we check if owner
attributes were registered (these consist of ID, UUID, PID and
name). The check then consists of checking if ID is not zero.
This is not going to work with
VIR_LOCK_MANAGER_OBJECT_TYPE_DAEMON which doesn't set ID. Switch
to setting PID which is available for both cases.

Signed-off-by: Michal Privoznik <mprivozn@redhat.com>
Message-Id: <68defdfe833ac39e4743353a34807265c42a5802.1535356707.git.mprivozn@redhat.com>