lock_driver: Introduce KEEP_OPEN flags
This flag causes connection to be opened when needed (e.g. when
calling virLockManagerLockDaemonAcquire for the first time) and
instead of closing it at the end of such API store it in
privateData so that it can be reused by later calls.

This is needed because if a resource is acquired and connection
is closed then virtlockd kills the registered PID (that's what
virtlockd is designed to do). Therefore we will need the
connection to open at drvAcquire and close not any sooner than
drvRelease. However, as we will be locking files step-by-step we
want to avoid opening new connection for every drvAcquire +
drvRelease pair, so the connection is going to be shared even
more than that. But more on that in next commit.

Signed-off-by: Michal Privoznik <mprivozn@redhat.com>
Message-Id: <cbc1abf5c982d734011b1e7d572555d514790d81.1535356707.git.mprivozn@redhat.com>