lock_manager: Introduce virLockManagerCloseConn
After the previous commit we have VIR_LOCK_MANAGER_ACQUIRE_KEEP_OPEN
flag. This is not enough because it will keep connection open for only
one instance of drvAcquire + drvRelease call. And when starting up a
domain there will be a lot of such calls as there will be a lot of paths
to relabel and thus lock. Therfore, VIR_LOCK_MANAGER_RELEASE_KEEP_OPEN
flag was introduced which allows us to keep connection open even after
the drvAcquire + drvRelease pair. In order to close the connection after
all locking has been done virLockManagerCloseConn is introduced.

Signed-off-by: Michal Privoznik <mprivozn@redhat.com>
Message-Id: <c7398caeb7dbb26b119468f9fd69ea3241892284.1535356707.git.mprivozn@redhat.com>